# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Share Link API",
    "version": "1.0"
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateServerConfig": {
        "type": "object",
        "properties": {
          "endpointUrl": {
            "type": "string",
            "description": "A string representing the external endpoint to retrieve data from.",
            "example": "https://example.com"
          },
          "clientId": {
            "type": "string",
            "required": false,
            "description": "A string representing a client's unique identifier for the external endpoint.",
            "example": null
          },
          "clientSecret": {
            "type": "string",
            "description": "A string representing a client's secret for the external endpoint.",
            "example": null,
            "required": false
          },
          "tokenEndpoint": {
            "type": "string",
            "required": false,
            "description": "An optional string representing the external endpoint to get a session token.",
            "example": "https://example.com"
          }
        }
      },
      "ServerConfig": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The user unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "endpointUrl": {
            "type": "string",
            "description": "A string representing the external endpoint to retrieve data from.",
            "example": "https://example.com"
          },
          "clientId": {
            "type": "string",
            "description": "A string representing a client's unique identifier for the external endpoint.",
            "example": "5AMl62z2XDmgrh2XsI2O"
          },
          "clientSecret": {
            "type": "string",
            "description": "A string representing a client's secret for the external endpoint.",
            "example": "5AMl62z2XDmgrh2XsI2O"
          },
          "tokenEndpoint": {
            "type": "string",
            "description": "An optional string representing the external endpoint to get a session token.",
            "example": "https://example.com"
          }
        }
      },
      "SHLinkAccessRequest": {
        "type": "object",
        "properties": {
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's management token.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          }
        }
      },
      "SHLinkAccess": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share link access' unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "shlinkId": {
            "type": "string",
            "description": "A string representing the unique identifier of share link.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "recipient": {
            "type": "string",
            "description": "A string representing the name of the viewer of the share link.",
            "example": "Jane Doe"
          },
          "accessTime": {
            "type": "datetime",
            "description": "A date representing the timestamp when the share link is accessed.",
            "example": "2024-09-09T09:45:00.479+00:00"
          }
        }
      },
      "CreateSHLinkEndpoint": {
        "type": "object",
        "properties": {
          "shlinkId": {
            "type": "string",
            "description": "A string representing the share link's unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's management token.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          }
        }
      },
      "SHLinkEndpoint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share link endpoint's unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "shlinkId": {
            "type": "string",
            "description": "A string representing the share link's unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's management token.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "serverConfigId": {
            "type": "string",
            "description": "A string representing the server config's unique identifier for the endpoint.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "urlPath": {
            "type": "string",
            "description": "An optional string representing the share link endpoint's query parameters.",
            "example": null
          }
        }
      },
      "SHLinkQRCodeRequest": {
        "type": "object",
        "properties": {
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's management token.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          }
        }
      },
...
      "SHLinkFile": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "description": "A string representing the content type of share link file.",
            "example": "application/smart-api-access"
          },
          "embedded": {
            "type": "string",
            "description": "An optional string representing the content of a share link's file.",
            "example": null
          },
          "location": {
            "type": "string",
            "description": "A string representing the share link's API link.",
            "example": "https://example.com/api/v1/shlinks/shlink-id/endpoints/endpoint-id?ticket=unique-ticket"
          }
        }
      },
...
      "SHLinkMini": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "A string representing the share links unique identifier.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "name": {
            "type": "string",
            "description": "A string representing the share link's name.",
            "example": "Jane Doe's Shlink"
          },
          "managementToken": {
            "type": "string",
            "description": "A string representing the share link's unique management token used for managing the share link.",
            "example": "hWWNwskdGOnEdq0KIQ3S"
          },
          "url": {
            "type": "string",
            "description": "A string representing the share link's encoded url.",
            "example": "5AMl62z2XDmgrh2XsI2O"
          },
          "passwordRequired": {
            "type": "boolean",
            "description": "A boolean indicating whether the share link has a passcode or not.",
            "example": true
          },
          "expiryDate": {
            "type": "datetime",
            "required": false,
            "description": "An optional date representing when the share link expires.",
            "example": "2024-09-09T09:45:00.479+00:00"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SHLinkFile"
            }
          }
        }
      },
...
    }
  },
  "security": [],
  "paths": {
    "/api/v1/server-configs": {
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Create a server config.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/CreateServerConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Create Server Config",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/ServerConfig"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Gets all server configs",
        "responses": {
          "200": {
            "description": "Server Configs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/ServerConfig"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shlinks/{id}/accesses": {
      "post": {
        "tags": [
          "Share Link Accesses"
        ],
        "description": "Get Share link Accesses.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/SHLinkAccessRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Share Link Accesses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SHLinkAccess"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shlinks/{id}/deactivate": {
      "delete": {
        "tags": [
          "Share Links"
        ],
        "description": "Deactivate a share link.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivated Share Link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/SHLinkMini"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shlinks/{id}/endpoints/{endpointId}": {
      "get": {
        "tags": [
          "Share Link Endpoints"
        ],
        "description": "Get a share link endpoint.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true
          },
          {
            "name": "endpointId",
            "in": "path",
            "description": "A string representing the share link endpoint's unique identifier.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get Share Link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shlinks/{id}/endpoints": {
      "post": {
        "tags": [
          "Share Link Endpoints"
        ],
        "description": "Create a share link endpoint.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "$ref": "#/components/schemas/CreateSHLinkEndpoint"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Get Share Link",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/SHLinkEndpoint"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/shlinks/{id}/qrcode": {
      "post": {
        "tags": [
          "Share Link QR Code"
        ],
        "description": "Get Share link QR Code as an image.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A string representing the share link's unique identifier.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SHLinkQRCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A PNG image of the share link's QR code.",
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}